chore: input validation 함수 추출
validation 관련 함수들을 `util/validateUserInfo`에 모아두었습니다
